use high quality audio
`high_quality` without stereo can also be used. but I think this one
provides better quality than that.

I tried setting up it as a run-time option, but I guess either you'll
have to restart the application or exit a group and re-join. So for now,
setting this as default.